# Conflicts:
#	configure
#	configure.in
#	include/mdsobjects.h
#	mdstcpip/Connections.c
#	mdstcpip/IoRoutinesTcp.c
#	mdstcpip/IoRoutinesUdt.c
#	mdstcpip/ProcessMessage.c
#	mdstcpip/mdsip_u.c
#	mdstcpip/mdsipshr.c

+++ camshr/compare_str.c
@@ -88,7 +88,7 @@ static int strncmp_nocase(const char *s, const char *t, int n)
   int i;
 
   if (MSGLVL(11))
-    printf("string sizes: *s(%d)  *t(%d)  n(%d)\n", strlen(s), strlen(t), n);
+    printf("string sizes: *s(%d)  *t(%d)  n(%d)\n", (int)strlen(s),(int)strlen(t), n);
 
   for (; n; n--, s++, t++) {
     if (MSGLVL(11))
